[ag/misc] Add global .agignore, and other misc updates
 - [vim] set nolinebreak for graphql files, since there's a lot of
 documentation
 - [setup] Add some todos for things to install
 - [setup] Don't run safelyLink without args

+++ safelyLink.sh
@@ -1,22 +1,32 @@
 #!/bin/bash
 
-file=$1
-target_dir=$2
-target_file_name=$(basename $file)
-target_file=$target_dir/$target_file_name
+function _f() {
+  file=$1
+  target_dir=$2
 
-if [ -e "$target_file" ]; then
-  # If it has no size (invalid or empty)
-  if [ ! -s "$target_file" ]; then
-    rm $target_file
-    # If it's a symlink
-  elif [ -L "$target_file" ]; then
-    echo "$target_file already linked, skipping"
-    exit
-  else
-    echo "$target_file exists, creating backup $target_file.bak"
-    mv $target_file $target_file.bak
+  if [ -z "$file" -o -z "$target_dir" ]; then
+    echo "No arguments provided, cancelling safelyLink"
+    return
   fi
-fi
 
-ln -sv $(pwd)/$file $target_dir
+  target_file_name=$(basename $file)
+  target_file=$target_dir/$target_file_name
+
+  if [ -e "$target_file" ]; then
+    # If it has no size (invalid or empty)
+    if [ ! -s "$target_file" ]; then
+      rm $target_file
+      # If it's a symlink
+    elif [ -L "$target_file" ]; then
+      echo "$target_file already linked, skipping"
+      exit
+    else
+      echo "$target_file exists, creating backup $target_file.bak"
+      mv $target_file $target_file.bak
+    fi
+  fi
+
+  ln -sv $(pwd)/$file $target_dir
+}
+
+_f "$@"
